+++ lib/glimmer/tk/frame_proxy.rb
@@ -33,6 +33,7 @@ module Glimmer
         if parent_proxy.is_a?(NotebookProxy)
           @tab_options, args[0] = args[0].to_h.partition {|key, value| NotebookProxy::TAB_OPTIONS.include?(key.to_s)}
           @tab_options = Hash[@tab_options]
+          @tab_options[:text] = @tab_options.delete(:title) if @tab_options.keys.include?(:title)
           args.delete_at(0) if args[0].to_a.empty?
         end
         super